NIFI-13609 Corrected Component Search Configuration
- Removed duplicate Controller Search Service bean definition
- Wired Controller Search Service to Controller Facade bean definition
